more reorganization of code
