+++ podaac/subsetter/subset.py
@@ -26,7 +26,7 @@ import geopandas as gpd
 import netCDF4 as nc
 import numpy as np
 import xarray as xr
-from shapely.geometry import MultiPolygon, Point, Polygon
+from shapely.geometry import Point
 
 from podaac.subsetter import (
     datatree_subset,
@@ -35,58 +35,12 @@ from podaac.subsetter import (
 from podaac.subsetter.utils import mask_utils
 from podaac.subsetter.utils import coordinate_utils
 from podaac.subsetter.utils import metadata_utils
+from podaac.subsetter.utils import spatial_utils
 from podaac.subsetter.utils import time_utils
 from podaac.subsetter.utils import file_utils
 
 
 SERVICE_NAME = 'l2ss-py'
-
-
-def translate_longitude(geometry):
-    """
-    Translates the longitude values of a Shapely geometry from the range [-180, 180) to [0, 360).
-
-    Parameters
-    ----------
-    geometry : shapely.geometry.base.BaseGeometry
-        The input shape geometry to be translated
-
-    Returns
-    -------
-    geometry
-        The translated shape geometry
-    """
-
-    def translate_point(point):
-        # Translate the point's x-coordinate (longitude) by adding 360
-        return Point((point.x + 360) % 360, point.y)
-
-    def translate_polygon(polygon):
-        def translate_coordinates(coords):
-            if len(coords[0]) == 2:
-                return [((x + 360) % 360, y) for x, y in coords]
-            if len(coords[0]) == 3:
-                return [((x + 360) % 360, y, z) for x, y, z in coords]
-            return coords
-
-        exterior = translate_coordinates(polygon.exterior.coords)
-
-        interiors = [
-            translate_coordinates(ring.coords)
-            for ring in polygon.interiors
-        ]
-
-        return Polygon(exterior, interiors)
-
-    if isinstance(geometry, (Point, Polygon)):  # pylint: disable=no-else-return
-        return translate_point(geometry) if isinstance(geometry, Point) else translate_polygon(geometry)
-    elif isinstance(geometry, MultiPolygon):
-        # Translate each polygon in the MultiPolygon
-        translated_polygons = [translate_longitude(subgeometry) for subgeometry in geometry.geoms]
-        return MultiPolygon(translated_polygons)
-    else:
-        # Handle other geometry types as needed
-        return geometry
 
 
 def subset_with_shapefile_multi(dataset: xr.Dataset,
@@ -132,7 +86,7 @@ def subset_with_shapefile_multi(dataset: xr.Dataset,
         # Convert shapefile to 0-360 if needed
         current_shapefile_df = shapefile_df.copy()
         if coordinate_utils.is_360(lon, lon_scale, lon_offset):
-            current_shapefile_df["geometry"] = current_shapefile_df["geometry"].apply(translate_longitude)
+            current_shapefile_df["geometry"] = current_shapefile_df["geometry"].apply(spatial_utils.translate_longitude)
 
         # Flatten points and convert to GeoDataFrame
         flat_points = np.column_stack((lon2d.ravel(), lat2d.ravel()))
