+++ podaac/subsetter/utils/spatial_utils.py
@@ -7,6 +7,7 @@ Utility functions for geospatial calculations and polygon operations.
 """
 
 import numpy as np
+from shapely.geometry import MultiPolygon, Point, Polygon
 from podaac.subsetter.utils import coordinate_utils
 
 
@@ -112,3 +113,50 @@ def get_east_west_lon(dataset, lon_var_name):
     eastmost = round(convert_to_standard(eastmost_360), 5)
     westmost = round(convert_to_standard(westmost_360), 5)
     return eastmost, westmost
+
+
+def translate_longitude(geometry):
+    """
+    Translates the longitude values of a Shapely geometry from the range [-180, 180) to [0, 360).
+
+    Parameters
+    ----------
+    geometry : shapely.geometry.base.BaseGeometry
+        The input shape geometry to be translated
+
+    Returns
+    -------
+    geometry
+        The translated shape geometry
+    """
+
+    def translate_point(point):
+        # Translate the point's x-coordinate (longitude) by adding 360
+        return Point((point.x + 360) % 360, point.y)
+
+    def translate_polygon(polygon):
+        def translate_coordinates(coords):
+            if len(coords[0]) == 2:
+                return [((x + 360) % 360, y) for x, y in coords]
+            if len(coords[0]) == 3:
+                return [((x + 360) % 360, y, z) for x, y, z in coords]
+            return coords
+
+        exterior = translate_coordinates(polygon.exterior.coords)
+
+        interiors = [
+            translate_coordinates(ring.coords)
+            for ring in polygon.interiors
+        ]
+
+        return Polygon(exterior, interiors)
+
+    if isinstance(geometry, (Point, Polygon)):  # pylint: disable=no-else-return
+        return translate_point(geometry) if isinstance(geometry, Point) else translate_polygon(geometry)
+    elif isinstance(geometry, MultiPolygon):
+        # Translate each polygon in the MultiPolygon
+        translated_polygons = [translate_longitude(subgeometry) for subgeometry in geometry.geoms]
+        return MultiPolygon(translated_polygons)
+    else:
+        # Handle other geometry types as needed
+        return geometry
